chore(build): move custom Vite plugins to vite/ directory
Extract all custom plugins and shared utilities out of the config files:

  vite/shared.js                              — logger, sharedOnwarn
  vite/plugins/inline-all-wasms.js            — WASM inlining plugin
  vite/plugins/rewrite-editor-worker-import.js — ?worker virtual modules
  vite/plugins/fix-cross-chunk-paths.js       — relative import rewriting

Remove vite.config.shared.js (superseded by vite/shared.js +
vite/plugins/inline-all-wasms.js). Complete migration of
vite.config.umd.js which still had local copies of all four utilities.
