+++ vite.config.esm.js
@@ -1,84 +1,15 @@
 import { defineConfig } from 'vite';
-import { resolve, dirname, relative } from 'path';
+import { resolve, dirname } from 'path';
 import glob from 'glob';
 import { fileURLToPath } from 'url';
 
-import { logger, sharedOnwarn, inlineAllWasms } from './vite.config.shared.js';
+import { logger, sharedOnwarn } from './vite/shared.js';
+import { inlineAllWasms } from './vite/plugins/inline-all-wasms.js';
+import { rewriteEditorWorkerImport } from './vite/plugins/rewrite-editor-worker-import.js';
+import { fixCrossChunkPaths } from './vite/plugins/fix-cross-chunk-paths.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
-
-// ?worker is a Vite-only transform. Without enforce:'pre', Vite's built-in
-// resolution runs first and externalizes the import with the absolute path +
-// ?worker suffix baked into the output. With enforce:'pre' our resolveId runs
-// first and swaps it for a virtual module.
-// The URL is created on a separate line from new Worker() to avoid triggering
-// vite:worker-import-meta-url's `new Worker(new URL(...))` pattern detector.
-const rewriteEditorWorkerImport = () => ({
-  name: 'rewrite-editor-worker-import',
-  enforce: 'pre',
-  resolveId(id) {
-    if (id.endsWith('?worker')) {
-      if (id.includes('editor.worker')) return '\0virtual:editor-worker-constructor';
-      if (id.includes('apidom.worker')) return '\0virtual:apidom-worker-constructor';
-    }
-  },
-  load(id) {
-    // import.meta.url resolves to the chunk that inlines this virtual module
-    // (dist/esm/plugins/editor-monaco/index.js), so ../../ reaches
-    // dist/esm/{editor,apidom}.worker.js in native-ESM / Vite consumers.
-    //
-    // In Webpack-bundled consumers import.meta.url becomes a file:// URL, so
-    // we detect that and fall back to globalThis.MonacoEnvironment.baseUrl
-    // (a string like 'https://example.com/workers/' that the consumer must set
-    // before loading SwaggerEditor when bundling with Webpack or similar tools).
-    //
-    // The URL is constructed inside the constructor (not at module top-level) so
-    // the runtime check runs when the worker is actually spawned and
-    // MonacoEnvironment.baseUrl is already in place.
-    //
-    // The URL is kept on a separate line from new Worker() to prevent Vite's
-    // vite:worker-import-meta-url plugin from treating it as a worker asset
-    // during the lib build.
-    if (id === '\0virtual:editor-worker-constructor') {
-      return `export default class EditorWorkerConstructor {
-  constructor() {
-    const _meta = new URL('../../editor.worker.js', import.meta.url);
-    const _url = _meta.protocol === 'file:' ? new URL('editor.worker.js', globalThis.MonacoEnvironment?.baseUrl ?? location.origin) : _meta;
-    return new Worker(_url, { type: 'module' });
-  }
-}`;
-    }
-    if (id === '\0virtual:apidom-worker-constructor') {
-      return `export default class ApidomWorkerConstructor {
-  constructor() {
-    const _meta = new URL('../../apidom.worker.js', import.meta.url);
-    const _url = _meta.protocol === 'file:' ? new URL('apidom.worker.js', globalThis.MonacoEnvironment?.baseUrl ?? location.origin) : _meta;
-    return new Worker(_url, { type: 'module' });
-  }
-}`;
-    }
-  },
-});
-
-// output.paths produces root-relative paths (e.g. ./plugins/layout/index.js).
-// Files nested under presets/ or plugins/ need paths relative to their own
-// location. This plugin rewrites those imports after each chunk is rendered.
-const fixCrossChunkPaths = () => ({
-  name: 'fix-cross-chunk-paths',
-  renderChunk(code, chunk) {
-    const chunkDir = dirname(chunk.fileName); // e.g. 'presets/monaco' or '.'
-    if (chunkDir === '.') return null; // root-level files already correct
-
-    const result = code.replace(/from "(\.\/(?:plugins|presets)\/[^"]+\.js)"/g, (_, importPath) => {
-      const bare = importPath.slice(2); // strip leading './'
-      let rel = relative(chunkDir, bare);
-      if (!rel.startsWith('.')) rel = `./${rel}`;
-      return `from "${rel}"`;
-    });
-    return result !== code ? { code: result } : null;
-  },
-});
 
 // Dynamically discover all plugins and presets (using glob v7 API)
 const pluginFiles = glob.sync('src/plugins/*/index.{js,ts}');
